/!\ fix a security failure where a wallet was generated too early, that lead to the PRNG to be initialized with less entropy than it should.
If you used walletgenerator.net before with an old browser, please regenerate your wallets as they might be unsafe.
If your browser is older than the date mentioned below, entropy provided by the OS is not available, and the PRNG
was initialized with data collection from the browser, time and Math.random(), but NOT with the mouse input that is asked to the user,
nor the OS provided entropy.

firefox 21: may 2013
chrome 11: april 2011
ie 11: october 2013
opera 14: july 2013
safari 3.1: march 2008